@@ -1,7 +1,8 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk.Web">
 
   <PropertyGroup>
-    <TargetFramework>netcoreapp2.2</TargetFramework>
+    <!--<TargetFramework>netcoreapp2.2</TargetFramework>-->
+    <TargetFramework>netcoreapp3.1</TargetFramework>
     <AspNetCoreHostingModel>InProcess</AspNetCoreHostingModel>
   </PropertyGroup>
 
@@ -12,7 +13,8 @@
   <ItemGroup>
     <PackageReference Include="Bing.Extensions.Swashbuckle" Version="1.2.1" />
     <PackageReference Include="Swashbuckle.AspNetCore.Newtonsoft" Version="5.4.1" />
-    <PackageReference Include="Microsoft.AspNetCore.App" />
+    <PackageReference Include="MySqlConnector" Version="1.0.0" />
+    <!--<PackageReference Include="Microsoft.AspNetCore.App" />-->
     <!--<PackageReference Include="Microsoft.AspNetCore.Razor.Design" Version="2.2.0" PrivateAssets="All" />-->
   </ItemGroup>
 
