@@ -38,7 +38,7 @@ namespace Bing.Admin.Modules
         /// <param name="services">服务集合</param>
         public override IServiceCollection AddServices(IServiceCollection services)
         {
-            services.AddSingleton<IConsumerServiceSelector, CapConsumerServiceSelector>();
+            //services.AddSingleton<IConsumerServiceSelector, CapConsumerServiceSelector>();
             LoadEvent(services);
             var config = services.GetConfiguration();
             var connection = config.GetConnectionString("DefaultConnection");
@@ -47,7 +47,7 @@ namespace Bing.Admin.Modules
             {
                 //o.UseEntityFramework<AdminUnitOfWork>();
                 o.UseMySql(connection);
-                o.UseDashboard();
+                //o.UseDashboard();
                 // 设置处理成功的数据在数据库中保存的时间（秒），为保证系统性能，数据会定期清理
                 o.SucceedMessageExpiredAfter = 24 * 3600;
                 // 设置失败重试次数
